Refactor solve test helpers into a new module
