+++ crates/compiler/solve/Cargo.toml
@@ -7,6 +7,11 @@ edition.workspace = true
 license.workspace = true
 version.workspace = true
 
+[[test]]
+name = "instantiate"
+path = "tests/instantiate.rs"
+harness = false
+
 [dependencies]
 roc_can = { path = "../can" }
 roc_collections = { path = "../collections" }
