+++ crates/compiler/solve/tests/instantiate.rs
@@ -1,0 +1,1 @@
+fn main() {}
